+++ hata/discord/http/http_client.py
@@ -1172,6 +1172,13 @@ class DiscordHTTPClient(HTTPClient):
             data,
         )
     
+    async def onboarding_screen_get(self, guild_id):
+        return await self.discord_request(
+            RateLimitHandler(RATE_LIMIT_GROUPS.onboarding_screen_get, guild_id),
+            METHOD_GET,
+            f'{API_ENDPOINT}/guilds/{guild_id}/onboarding',
+        )
+    
     async def voice_state_client_edit(self, guild_id, data):
         return await self.discord_request(
             RateLimitHandler(RATE_LIMIT_GROUPS.voice_state_client_edit, NO_SPECIFIC_RATE_LIMITER),